[FIX] Refactor controller/notifications
