+++ lib/settlements.js
@@ -19,6 +19,8 @@ const NoRelatedDestinationDebitError =
   require('../errors/no-related-destination-debit-error')
 const ManyToManyNotSupportedError =
   require('../errors/many-to-many-not-supported-error')
+const UnrelatedNotificationError =
+  require('../errors/unrelated-notification-error')
 const AssetsNotTradedError = require('../errors/assets-not-traded-error')
 const hashJSON = require('@ripple/five-bells-shared/utils/hashJson')
 
@@ -526,16 +528,38 @@ Settlements.prototype.settle = function * (settlement) {
   }
 }
 
+Settlements.prototype.updateTransfer = function * (updatedTransfer) {
+  // Maybe its a source transfer:
+  let settlement = this.sourceSubscriptions.get(updatedTransfer.id)
+  if (settlement) {
+    yield this.updateSourceTransfer(updatedTransfer, settlement)
+    return
+  }
+
+  // Or a destination transfer:
+  let sourceTransfers = this.destinationSubscriptions.get(updatedTransfer.id)
+  if (sourceTransfers && sourceTransfers.length) {
+    yield this.updateDestinationTransfer(updatedTransfer, sourceTransfers)
+    return
+  }
+
+  // TODO: should we delete the subscription?
+  throw new UnrelatedNotificationError('Notification does not match a ' +
+    'settlement we have a record of or the corresponding source ' +
+    'transfers may already have been executed')
+}
+
 // A notification about `updatedTransfer` (which is a source transfer) arrived,
 // so update the state on the corresponding transfer within the cached
 // settlement. When all of the settlement's source transfers are "prepared",
 // authorized/submit the payment.
-Settlements.prototype.updateSourceTransfer = function * (settlement, updatedTransfer) {
+Settlements.prototype.updateSourceTransfer = function * (updatedTransfer, settlement) {
   let target_id = updatedTransfer.id
   let source_transfers = settlement.source_transfers
   for (var i = 0; i < source_transfers.length; i++) {
     if (source_transfers[i].id === target_id) {
       source_transfers[i] = updatedTransfer
+      log.debug('updateSourceTransfer', target_id)
       break
     }
   }
@@ -545,6 +569,26 @@ Settlements.prototype.updateSourceTransfer = function * (settlement, updatedTran
     if (!this.sourceSubscriptions.hasSettlement(settlement)) {
       yield this.settle(settlement)
     }
+  }
+}
+
+Settlements.prototype.updateDestinationTransfer = function * (updatedTransfer, sourceTransfers) {
+  if (updatedTransfer.state !== 'executed') {
+    log.debug('got notification about unknown or incomplete transfer')
+    return
+  }
+
+  // TODO: make sure the transfer is signed by the ledger
+  log.debug('got notification about executed destination transfer')
+
+  // This modifies the source_transfers states
+  yield executeSourceTransfers(sourceTransfers, [updatedTransfer])
+  let allTransfersExecuted = _.every(sourceTransfers, function (transfer) {
+    return transfer.state === 'executed'
+  })
+  if (!allTransfersExecuted) {
+    log.error('not all source transfers have been executed, ' +
+      'meaning we have not been fully repaid')
   }
 }
